Started to re-arrange loader functionality.
I'm currently looking at creating loaderFactories that implement an interface for each tab.  Still trying to figure out whether I should make it 1 factory per tab, or 1 factory per application (cycling, xc, etc).  The loaders have real implementations for what to retrieve from the database, so these could be swapped out with the database to change the model of the application.  The UI would just have to bind correctly to the cursors that are being returned, so they would have to know what they are getting back in the cursors, but wouldn't have to know anything about where they got it from.